Update code for 6.8

+++ src/cse/bubt/edu/bd/Chapter_6/_601_Max_Three_Number.java
@@ -6,10 +6,10 @@ public class _601_Max_Three_Number {
     public static void main(String[] args) {
         Scanner input=new Scanner(System.in);
         double a,b,c;
-        a=input.nextInt();
-        b=input.nextInt();
-        c=input.nextInt();
-        double d=greater(a,b,c);
+        a=input.nextDouble();
+        b=input.nextDouble();
+        c=input.nextDouble();
+        double d = greater(a,b,c);
         System.out.println(d);
     }
     public static double greater(double a, double b, double c) {
